backend/requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
SQLAlchemy==2.0.10
pymysql==1.0.2
python-dotenv==1.0.0
pydantic[email]
redis==4.5.5